+++ packages/vitest/src/utils/index.ts
@@ -61,44 +61,53 @@ export async function ensurePackageInstalled(
   return false
 }
 
-/**
- * Deep merge :P
- */
-export function deepMerge<T extends object = object>(target: T, ...sources: any[]): T {
-  if (!sources.length)
-    return target as any
+export function isObject(item: unknown): boolean {
+  return item != null && typeof item === 'object' && !Array.isArray(item)
+}
 
-  const source = sources.shift()
-  if (source === undefined)
-    return target as any
+function deepMergeArray(target: any[] = [], source: any[] = []) {
+  const mergedOutput = Array.from(target)
 
-  if (isMergableObject(target) && isMergableObject(source)) {
+  source.forEach((sourceElement, index) => {
+    const targetElement = mergedOutput[index]
+
+    if (Array.isArray(target[index])) {
+      mergedOutput[index] = deepMergeArray(target[index], sourceElement)
+    }
+    else if (isObject(targetElement)) {
+      mergedOutput[index] = deepMerge(target[index], sourceElement)
+    }
+    else {
+      // Source does not exist in target or target is primitive and cannot be deep merged
+      mergedOutput[index] = sourceElement
+    }
+  })
+
+  return mergedOutput
+}
+
+export function deepMerge(target: any, source: any): any {
+  if (isObject(target) && isObject(source)) {
+    const mergedOutput = { ...target }
     Object.keys(source).forEach((key) => {
-      if (isMergableObject(source[key])) {
-        // @ts-expect-error
-        if (!target[key])
-          // @ts-expect-error
-          target[key] = {}
-
-        // @ts-expect-error
-        deepMerge(target[key], source[key])
+      if (isObject(source[key]) && !source[key].$$typeof) {
+        if (!(key in target)) Object.assign(mergedOutput, { [key]: source[key] })
+        else mergedOutput[key] = deepMerge(target[key], source[key])
+      }
+      else if (Array.isArray(source[key])) {
+        mergedOutput[key] = deepMergeArray(target[key], source[key])
       }
       else {
-        // @ts-expect-error
-        target[key] = source[key]
+        Object.assign(mergedOutput, { [key]: source[key] })
       }
     })
+
+    return mergedOutput
   }
-
-  return deepMerge(target, ...sources)
-}
-
-function isMergableObject(item: any): item is Object {
-  return isObject(item) && !Array.isArray(item)
-}
-
-export function isObject(val: any): val is object {
-  return toString.call(val) === '[object Object]'
+  else if (Array.isArray(target) && Array.isArray(source)) {
+    return deepMergeArray(target, source)
+  }
+  return target
 }
 
 export function toFilePath(id: string, root: string): string {
